Add GET /note/search endpoint and dependent code

+++ notes-app/src/main/java/pl/michallysak/notes/note/model/PagedQuery.java
@@ -2,7 +2,7 @@ package pl.michallysak.notes.note.model;
 
 public interface PagedQuery {
 
-  long getPage();
+  int getPage();
 
-  long getSize();
+  int getSize();
 }
